refactor(tasks): migrate train module Signed-off-by: Cameron Smith <cameron.ray.smith@gmail.com>
