+++ src/pyrovelocity/interfaces.py
@@ -10,7 +10,7 @@ from mashumaro.mixins.json import DataClassJSONMixin
 
 from pyrovelocity.tasks.data import download_dataset
 from pyrovelocity.tasks.preprocess import preprocess_dataset
-from pyrovelocity.train import train_dataset
+from pyrovelocity.tasks.train import train_dataset
 from pyrovelocity.workflows.configuration import create_dataclass_from_callable
