+++ src/pyrovelocity/tasks/__init__.py
@@ -1,7 +1,10 @@
 import pyrovelocity.tasks.data
 import pyrovelocity.tasks.preprocess
+import pyrovelocity.tasks.train
+
 
 __all__ = [
     "data",
     "preprocess",
+    "train",
 ]
